feat(StatsGraphs): Update player production at end of step
